[#89] refactor: formatDate 공통 유틸리티로 추출
